+++ src/utils/filters/index.ts
@@ -1,0 +1,4 @@
+export {
+  getColumnSearchFilterProps,
+  getColumnTreeSelectFilterProps,
+} from "./tableFilters";
